backend/pyproject.toml
[project]
name = "genslides-backend"
version = "0.1.0"
description = "GenSlides Backend API - AI-powered slide generation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "pyyaml>=6.0.3",
    "blake3>=1.0.8",
    "google-genai>=1.59.0",
    "pillow>=12.1.0",
    "websockets>=15.0.1",
    "httpx>=0.28.1",
    "aiofiles>=25.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
    "mypy>=1.19.1",
    "types-aiofiles>=24.1.0",
    "types-pyyaml>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
